[#4] Feat: 메인화면 사이즈에 따른 이미지 변경 구현
